Removendo static files
