--- backend/project/apps/home/templates/home.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-  <head>
-    {% include "partials/_head.html" %}
-  </head>
-  <body>
-    <div class="content-block">
-      <div class="form-wrapper">
-        <h1>Bem vindo a plataforma</h1>
-        <div  class="account-link">
-          <h2><a href="{% url "user:login" %}">Vamos começar?</a></h2>
-        </div>
-      </div>
-    </div>
-  </body>
-</html>
