--- backend/project/apps/home/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from project.apps.home import views
-
-app_name = "home"
-
-urlpatterns = [
-    path("", views.home, name="home"),
-]
