--- backend/project/apps/home/views.py
@@ -1,8 +1,0 @@
-from django.shortcuts import render, redirect
-
-
-def home(request):
-    user = request.user
-    if user.is_authenticated:
-        return redirect("question:index")
-    return render(request, "home.html", {"site_title": "Home"})
