--- backend/project/apps/question/templates/index.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %}
-
-{% block 'questions' %}
-{% include "partials/_header.html" %}
-
-<div class="questions-block">
-  {% if not questions %}
-    <div class="global-title"><h1>Nenhuma questão encontrada :(</h1></div>
-  {% else %}
-    <div class="posts-wrapper">
-      {% for question in questions %}
-        {% if question.user == request.user %}
-          <div class="post-box">
-            <a href="{% url "question:question" question.id %}" class="card-link">
-              {% include "partials/question_card.html" with i=forloop.counter0 %}
-            </a>
-          </div>
-        {% endif %}
-      {% endfor %}
-    </div>
-  {% endif %}
-</div>
-{% endblock 'questions' %}
