--- backend/project/apps/question/templates/partials/question_card.html
@@ -1,10 +1,0 @@
-<article class="card">
-  <div class="card-title">
-    <p>{{ question.theme }} | {{ question.edition }}</p>
-  </div>
-  <div class="card-content">
-    <p>
-        {{ question }}
-    </p>
-  </div>
-</article>
