+++ backend/project/apps/question/urls.py
@@ -1,18 +1,11 @@
 from django.urls import path
-from project.apps.question import views
-from project.apps.question.api.views import getQuestions, getQuestion
+
+from project.apps.question.views import getQuestions, getQuestion
 
 app_name = "question"
 
 urlpatterns = [
-    path("index/", views.index, name="index"),
-    path("question/<int:question_id>/", views.question, name="question"),
-    path(
-        "question/<int:question_id>/result",
-        views.answer_question,
-        name="answer_question",
-    ),
     # API endpoints
-    path("api/questions/", getQuestions, name="questions"),
-    path("api/questions/<int:id>/", getQuestion, name="question_detail"),
+    path("", getQuestions, name="questions"),
+    path("<int:id>/", getQuestion, name="question_detail"),
 ]
