--- backend/project/apps/user/templates/login.html
@@ -1,20 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="content-block">
-  <div class="form-wrapper">
-    <h1>Entre com sua conta</h1>
-      {% if messages %}
-        {% for message in messages %}
-          <div class="{{ message.tags }}-message {{ message.tags }}-box">
-            {{ message }}
-          </div>
-        {% endfor %}
-      {% endif %}
-    {% include "partials/user_register.html" %}
-    <div class="account-link">
-      <p>Não tem conta? <a href="{% url "user:register" %}">Crie sua conta aqui</a></p>
-    </div>
-  </div>
-</div>
-{% endblock content %}
