--- backend/project/apps/user/templates/partials/user_register.html
@@ -1,30 +1,0 @@
-<form action="{{ form_action }}" method="POST">
-  {% csrf_token %}
-  <div>
-    {% for field in form %}
-    <div class="input-box">
-      <label for="{{ field.id_for_label }}" class="input-label"
-      >
-      {{ field }}
-      {% if field.errors %}
-        {% for error in field.errors %}
-          <li class="error-message"> {{error}} </li>
-        {% endfor %}
-      {% endif %}
-
-      {% if field.help_text %}
-        {% for help_text in field.help_text  %}
-          <li class="help-text">{{ help_text }}</li>
-        {% endfor %} 
-      {% endif %}
-    </div>
-    {% endfor %}
-  </div>
-  <div>
-    {% if site_title == "Crie sua conta" %}
-      <p><button class="btn-confirm" type="submit">Concluir</button></p>
-    {% else %}
-      <p><button class="btn-confirm" type="submit">Entrar</button></p>
-    {% endif %}
-  </div>
-</form>
