--- backend/project/apps/user/templates/register.html
@@ -1,13 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="content-block">
-  <div class="form-wrapper">
-    <h1>Crie sua conta</h1>
-    {% include "partials/user_register.html" %}
-    <div class="account-link">
-      <p>Já tem uma conta? <a href="{% url "user:login" %}">Faça login por aqui</a></p>
-    </div>
-  </div>
-</div>
-{% endblock content %}
