--- backend/project/templates/base.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-  <head>
-    {% include "partials/_head.html" %}
-  </head>
-  <body> 
-    {% block content %}{% endblock content %} 
-    {% block 'questions'%}{% endblock 'questions' %}
-    {% block 'single_question' %}{% endblock 'single_question' %}
-
-    {% block options %}{% endblock options %}
-    
-  </body>
-</html>
