--- backend/project/templates/partials/_head.html
@@ -1,13 +1,0 @@
-{% load static %}
-<meta charset="UTF-8" />
-<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-<link rel="preconnect" href="https://fonts.googleapis.com" />
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-<link
-  href="https://fonts.googleapis.com/css2?family=Ancizar+Sans:ital,wght@0,100..1000;1,100..1000&display=swap"
-  rel="stylesheet"
-/>
-
-<link rel="stylesheet" href="{% static "css/style.css" %}">
-
-<title>{{site_title}}</title>
