--- backend/project/templates/partials/_header.html
@@ -1,44 +1,0 @@
-{% load static %}
-
-<header class="menu-wrapper">
-  <nav class="menu-bar"> 
-    <div class="menu-logo">
-      <a href="#">
-        <img src="{% static "img/book-icon.svg" %}" alt="">
-      </a>
-    </div>
-    <div class="menu-list">
-      <ul>
-        <li class="menu-item"><a href="{% url "question:index" %}">Início</a></li>
-      </ul>
-    </div>
-    <div class="menu-icon">
-      <button onclick="menuOnClick()">
-        <img class="icon" src="{% static 'img/menu-icon.svg' %}" alt="">
-        <script>
-          function menuOnClick() {
-            let altMenu = document.querySelector(".alt-menu");
-
-            if (altMenu.classList.contains("open")) {
-              altMenu.classList.remove("open");
-              document.querySelector(".icon").src = "{% static 'img/menu-icon.svg' %}";
-            } else {
-              altMenu.classList.add("open");
-              document.querySelector(".icon").src = "{% static 'img/close-icon.svg' %}";
-            }
-          }
-        </script>
-      </button>
-    </div>
-  </nav>
-
-  <div class="alt-menu">
-    <div class="div-line">
-      <ul>
-        <li class="menu-item"><a href="{% url "user:add_questions" %}">Adicionar questões</a></li>
-        <li class="menu-item"><a href="{% url "user:logout" %}">Logout</a></li>
-      </ul>
-    </div>
-  </div>
-</header>
-
